POC-1/requirements.txt
fastapi>=0.95.1
uvicorn[standard]>=0.22.0
python-dotenv>=0.21.0
wsdiscovery>=0.6.3
onvif-zeep>=0.1.13
pydantic>=1.10.2